CI changes

+++ cpp/models/ode_secir/model.h
@@ -395,8 +395,8 @@ IOResult<ScalarType> get_reproduction_number(size_t t_idx, const Simulation<Base
 
         for (Eigen::Index l = 0; l < (Eigen::Index)num_groups; l++) {
             cont_freq_eff(l, (size_t)k) =
-                season_val * contact_matrix.get_matrix_at(t_idx)(static_cast<Eigen::Index>((size_t)l),
-                                                                 static_cast<Eigen::Index>((size_t)k));
+                season_val * contact_matrix.get_matrix_at(static_cast<double>(t_idx))(
+                                 static_cast<Eigen::Index>((size_t)l), static_cast<Eigen::Index>((size_t)k));
         }
     }
